#19 Refactoring JobList panel features to SlidePanel
